\name{callArg}
\alias{callArg}
\title{Work with plot call arguments}
\description{
  Part of the \code{\link{playwith}} Application Programming Interface.
}
\usage{
callArg(playState, arg, name = NULL)
callArg(playState, arg, name = NULL) <- value
}
\arguments{
  \item{playState}{ a \code{\link{playState}} object representing the plot, window and device. }
  \item{arg}{  }
  \item{name}{  }
  \item{value}{  }
}
\details{
  
}
\value{
	
}
\author{ Felix Andrews \email{felix@nfrac.org}  }
\seealso{ \code{\link{playwith.API}} }
\examples{
}
\keyword{ programming }

