\name{playDev}
\alias{playDev}
\alias{playDevCur}
\alias{playDevList}
\alias{playDevSet}
\alias{playDevOff}
\title{Work with the playwith device}
\description{
  Part of the \code{\link{playwith}} Application Programming Interface.
}
\usage{
playDevCur()
playDevList()
playDevSet(playState)
playDevOff(playState = playDevCur())
}
\arguments{
  \item{playState}{ a \code{\link{playState}} object representing the plot, window and device. }
}
\details{
  
}
\value{
	
}
\author{ Felix Andrews \email{felix@nfrac.org}  }
\seealso{ \code{\link{playwith.API}} }
\examples{
}
\keyword{ programming }

