\name{playDo}
\alias{playDo}
\title{Do something in a specified plot space}
\description{
  Part of the \code{\link{playwith}} Application Programming Interface.
}
\usage{
playDo(playState, expr, space = "plot", clip.off = FALSE)
}
\arguments{
  \item{playState}{ a \code{\link{playState}} object representing the plot, window and device. }
  \item{expr}{  }
  \item{space}{  }
  \item{clip.off}{  }
}
\details{
  
}
\value{
	
}
\author{ Felix Andrews \email{felix@nfrac.org}  }
\seealso{ \code{\link{playwith.API}} }
\examples{
}
\keyword{ programming }

