\name{playPrompt}
\alias{playPrompt}
\alias{playFreezeGUI}
\alias{playThawGUI}
\alias{blockRedraws}
\title{Show a message to the playwith user}
\description{
  Part of the \code{\link{playwith}} Application Programming Interface.
}
\usage{
playPrompt(playState, text = NULL)
playFreezeGUI(playState)
playThawGUI(playState)
blockRedraws(expr, playState = playDevCur())
}
\arguments{
  \item{playState}{ a \code{\link{playState}} object representing the plot, window and device. }
  \item{text}{  }
  \item{expr}{  }
}
\details{
  
}
\value{
	
}
\author{ Felix Andrews \email{felix@nfrac.org}  }
\seealso{ \code{\link{playwith.API}} }
\examples{
}
\keyword{ programming }

