\name{playTools}
\alias{playTools}
\alias{toolConstructors}
\alias{playApplicationTools}
\alias{playInteractionTools}
\alias{play3DTools}
\alias{playSplomTools}
\title{Built-in tools for playwith}
\description{
  This document describes the built-in tools that can be placed in 
  a \code{\link{playwith}} window. The constructor functions themselves
  can be found in the named list \code{toolConstructors}.
}
\details{
	Tools in the set \code{playApplicationTools}, by default on the top toolbar:
	\describe{
		\item{\code{options}}{
		}
		\item{\code{stayontop}}{
		}
		\item{\code{keep}}{
		}
		\item{\code{save}}{
		}
		\item{\code{copy}}{
		}
		\item{\code{print}}{
		}
		\item{\code{data}}{
		}
		\item{\code{settings}}{
		}
		\item{\code{theme}}{
		}
		\item{\code{time.mode}}{
		}
	}
	
	Tools in the set \code{playInteractionTools}, by default on the left toolbar:
	\describe{
		\item{\code{expand}}{
		}
		\item{\code{identify}}{
		}
		\item{\code{annotate}}{
		}
		\item{\code{arrow}}{
		}
		\item{\code{edit.annotations}}{
		}
		\item{\code{clear}}{
		}
		\item{\code{zoom}}{
		}
		\item{\code{zoomout}}{
		}
		\item{\code{zoomfit}}{
		}
		\item{\code{zero}}{
		}
		\item{\code{coords}}{
		}
	}
	
	Tools in the set \code{play3DTools}:
	\describe{
		\item{\code{zoomin.3d}}{
		}
		\item{\code{zoomout.3d}}{
		}
		\item{\code{fly.left.3d}}{
		}
		\item{\code{fly.right.3d}}{
		}
	}
	
	Tools in the set \code{playSplomTools}:
	\describe{
		\item{\code{brush}}{
		}
	}

}
\author{ Felix Andrews \email{felix@nfrac.org} }
\seealso{ \code{\link{playwith}}, \link{playwith.API} }
\examples{
	
}
\keyword{ iplot }
\keyword{ dynamic }


