\name{whichSpace}
\alias{whichSpace}
\alias{deviceCoordsToSpace}
\title{Convert device coordinates to plot space}
\description{
  Part of the \code{\link{playwith}} Application Programming Interface.
}
\usage{
whichSpace(playState, x.device, y.device)

deviceCoordsToSpace(playState, x.device, y.device, space = "plot")
}
\arguments{
  \item{playState}{ a \code{\link{playState}} object representing the plot, window and device. }
  \item{x.device}{  }
  \item{y.device}{  }
  \item{space}{  }
}
\details{
  
}
\value{
	
}
\author{ Felix Andrews \email{felix@nfrac.org}  }
\seealso{ \code{\link{playwith.API}} }
\examples{
}
\keyword{ programming }

