\name{xyCoords}
\alias{xyCoords}
\alias{xyData}
\title{Get playwith plot data points}
\description{
  Part of the \code{\link{playwith}} Application Programming Interface.
}
\usage{
xyCoords(playState, space = "plot")
xyData(playState, space = "plot")
}
\arguments{
  \item{playState}{ a \code{\link{playState}} object representing the plot, window and device. }
  \item{space}{  }
}
\details{
  
}
\value{
	
}
\author{ Felix Andrews \email{felix@nfrac.org}  }
\seealso{ \code{\link{playwith.API}} }
\examples{
}
\keyword{ programming }

