\name{playReplot}
\alias{playReplot}
\alias{playNewPlot}
\title{Draw a playwith plot}
\description{
  Part of the \code{\link{playwith}} Application Programming Interface.
}
\usage{
playNewPlot(playState)
playReplot(playState)
}
\arguments{
  \item{playState}{ a \code{\link{playState}} object representing the plot, window and device. }
}
\details{
  
}
\value{
	
}
\author{ Felix Andrews \email{felix@nfrac.org}  }
\seealso{ \code{\link{playwith.API}} }
\examples{
}
\keyword{ programming }

