\name{playSelectData}
\alias{playSelectData}
\alias{playPointInput}
\alias{playLineInput}
\alias{playRectInput}
\title{Get input from the playwith user}
\description{
  Part of the \code{\link{playwith}} Application Programming Interface.
}
\usage{
playSelectData(playState, prompt = "Click or drag to select data points.")
playPointInput(playState, prompt = "Click on the plot")
playLineInput(playState, prompt = "Click and drag to define a line")
playRectInput(playState, prompt = "Click and drag to define a rectangular region")
}
\arguments{
  \item{playState}{ a \code{\link{playState}} object representing the plot, window and device. }
  \item{prompt}{  }
}
\details{
  
}
\value{
	
}
\author{ Felix Andrews \email{felix@nfrac.org}  }
\seealso{ \code{\link{playwith.API}} }
\examples{
}
\keyword{ programming }

