\name{plotadd}
\alias{plotadd}
\title{ Augment a plot in-line }
\description{
  A simple way to encapsulate a plot and annotations in one call. 
}
\usage{
plotadd(FUN, ..., add.stuff = expression())
}
\arguments{
  \item{FUN}{ a high-level plot function. }
  \item{\dots}{ arguments to \code{FUN}. }
  \item{add.stuff}{ expressions to augment the plot, evaluated after plotting. }
}
\details{
  
}
\value{
  
}
\author{ Felix Andrews \email{felix@nfrac.org} }
\examples{

}
\keyword{ hplot }
\keyword{ aplot }

